pyproject.toml
[project]
name = "fastapi-temp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.3"
dependencies = [
    "alembic>=1.16.1",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "email-validator>=2.2.0",
    "emails>=0.6",
    "fastapi-limiter>=0.1.6",
    "fastapi-mail>=1.5.0",
    "fastapi[all]>=0.115.12",
    "httpx[http2]>=0.28.1",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.6",
    "passlib>=1.7.4",
    "pydantic-settings>=2.9.1",
    "pydantic>=2.11.5",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "redis>=6.2.0",
    "rich>=14.0.0",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.41",
    "sqlmodel>=0.0.24",
    "tenacity>=9.1.2",
    "argon2-cffi>=23.1.0",
    "premailer>=3.10.0",
]

[tool.uv]
dev-dependencies = [
    "pytest<8.0.0,>=7.4.3",
    "mypy<2.0.0,>=1.8.0",
    "ruff<1.0.0,>=0.2.2",
    "pre-commit<4.0.0,>=3.6.2",
    "types-passlib<2.0.0.0,>=1.7.7.20240106",
    "coverage<8.0.0,>=7.4.3",
]
